# core: Remove await from write in http_bench